Load/edit/retrieve items from database.
I'm having this issue where I have a race condition between getting the siteInfo/userInfo, and mounting Vue pages. I've been able to solve it by manually querying for userInfo once the Vue page has been mounted, then calling methods that require them, but it feels quite hacky.

Other sites don't seem to have to deal with this... Basically I want it to grab the siteInfo before mounting any pages or anything so that it's all available for when I call it.